+++ src_docs/source/releases.rst
@@ -6,6 +6,8 @@ Releases
 v0.9.0
 ------
 * Added function to test push targets from configuration page. It will send data and show the return code as a first step. 
+* Merged index and device pages into one so that all the needed information is available on the index page.
+* Removed api for device (/api/device), it's now merged into the /api/status api.
 * Updated WifiManager and DoubleReset libraries
 * Experimental release of firmware using an esp32 instead of esp8266
 * Updated esp32 target with littlefs support